requirements.txt
aliyun-log-python-sdk==0.9.11
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altgraph==0.17.4
async-timeout==4.0.3
attrs==24.3.0
branca==0.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
crcmod==1.7
cryptography==43.0.1
dateparser==1.2.0
elastic-transport==8.15.0
elasticsearch==8.15.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
gevent==24.2.1
greenlet==3.1.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==0.10.0
jsonpath==0.82.2
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
oss2==2.19.0
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
pluggy==1.5.0
protobuf==5.28.2
pycparser==2.22
pycryptodome==3.20.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pynmea2==1.19.0
pyserial==3.5
pytest==8.3.3
pytest-ordering==0.6
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
regex==2024.9.11
requests==2.32.3
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
websocket==0.2.1
xlrd==2.0.1
XlsxWriter==3.2.0
xlutils==2.0.0
xlwt==1.3.0
xyzservices==2024.9.0
bleak==0.22.2